119061 refactor observability (#119211)
* group containers and components, create index files
# Conflicts:
#	x-pack/plugins/observability/public/pages/alerts/containers/alerts_table_t_grid/alerts_table_t_grid.tsx
#	x-pack/plugins/observability/public/pages/cases/helpers.ts
#	x-pack/plugins/observability/public/routes/index.tsx
