+++ x-pack/plugins/observability/public/components/app/cases/case_view/helpers.ts
@@ -8,7 +8,7 @@ import { useEffect, useState } from 'react';
 import { isEmpty } from 'lodash';
 
 import { usePluginContext } from '../../../../hooks/use_plugin_context';
-import { parseAlert } from '../../../../pages/alerts/parse_alert';
+import { parseAlert } from '../../../../pages/alerts/components/parse_alert';
 import { TopAlert } from '../../../../pages/alerts/';
 import { useKibana } from '../../../../utils/kibana_react';
 import { Ecs } from '../../../../../../cases/common';
