+++ x-pack/plugins/observability/public/index.ts
@@ -64,7 +64,9 @@ export {
   METRIC_TYPE,
 } from './hooks/use_track_metric';
 
-export const LazyAlertsFlyout = lazy(() => import('./pages/alerts/alerts_flyout'));
+export const LazyAlertsFlyout = lazy(
+  () => import('./pages/alerts/components/alerts_flyout/alerts_flyout')
+);
 export { useFetcher, FETCH_STATUS } from './hooks/use_fetcher';
 export { useEsSearch, createEsParams } from './hooks/use_es_search';
