+++ x-pack/plugins/observability/public/pages/alerts/components/alerts_flyout/alerts_flyout.test.tsx
@@ -6,11 +6,11 @@
  */
 
 import React from 'react';
-import * as useUiSettingHook from '../../../../../../../src/plugins/kibana_react/public/ui_settings/use_ui_setting';
-import { createObservabilityRuleTypeRegistryMock } from '../../../rules/observability_rule_type_registry_mock';
-import { render } from '../../../utils/test_helper';
-import type { TopAlert } from '../';
-import { AlertsFlyout } from './';
+import * as useUiSettingHook from '../../../../../../../../src/plugins/kibana_react/public/ui_settings/use_ui_setting';
+import { createObservabilityRuleTypeRegistryMock } from '../../../../rules/observability_rule_type_registry_mock';
+import { render } from '../../../../utils/test_helper';
+import type { TopAlert } from '../../containers/alerts_page';
+import { AlertsFlyout } from '..';
 
 describe('AlertsFlyout', () => {
   jest
