+++ x-pack/plugins/observability/public/pages/alerts/components/alerts_flyout/alerts_flyout.tsx
@@ -35,14 +35,14 @@ import {
 } from '@kbn/rule-data-utils/alerts_as_data_status';
 import moment from 'moment-timezone';
 import React, { useMemo } from 'react';
-import type { TopAlert } from '../';
-import { useKibana, useUiSetting } from '../../../../../../../src/plugins/kibana_react/public';
-import { asDuration } from '../../../../common/utils/formatters';
-import type { ObservabilityRuleTypeRegistry } from '../../../rules/create_observability_rule_type_registry';
+import type { TopAlert } from '../../containers';
+import { useKibana, useUiSetting } from '../../../../../../../../src/plugins/kibana_react/public';
+import { asDuration } from '../../../../../common/utils/formatters';
+import type { ObservabilityRuleTypeRegistry } from '../../../../rules/create_observability_rule_type_registry';
 import { parseAlert } from '../parse_alert';
-import { AlertStatusIndicator } from '../../../components/shared/alert_status_indicator';
-import { ExperimentalBadge } from '../../../components/shared/experimental_badge';
-import { translations, paths } from '../../../config';
+import { AlertStatusIndicator } from '../../../../components/shared/alert_status_indicator';
+import { ExperimentalBadge } from '../../../../components/shared/experimental_badge';
+import { translations, paths } from '../../../../config';
 
 type AlertsFlyoutProps = {
   alert?: TopAlert;
