+++ x-pack/plugins/observability/public/pages/alerts/components/alerts_flyout/example_data.ts
