+++ x-pack/plugins/observability/public/pages/alerts/components/default_cell_actions.tsx
@@ -9,9 +9,9 @@ import React from 'react';
 import { i18n } from '@kbn/i18n';
 import { getMappedNonEcsValue } from './render_cell_value';
 import FilterForValueButton from './filter_for_value';
-import { TimelineNonEcsData } from '../../../../timelines/common/search_strategy';
-import { TGridCellAction } from '../../../../timelines/common/types/timeline';
-import { getPageRowIndex } from '../../../../timelines/public';
+import { TimelineNonEcsData } from '../../../../../timelines/common/search_strategy';
+import { TGridCellAction } from '../../../../../timelines/common/types/timeline';
+import { getPageRowIndex } from '../../../../../timelines/public';
 
 export const FILTER_FOR_VALUE = i18n.translate('xpack.observability.hoverActions.filterForValue', {
   defaultMessage: 'Filter for value',
