+++ x-pack/plugins/observability/public/pages/alerts/components/render_cell_value/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { getRenderCellValue, getMappedNonEcsValue } from './render_cell_value';
