+++ x-pack/plugins/observability/public/pages/alerts/components/severity_badge/severity_badge.stories.tsx
