+++ x-pack/plugins/observability/public/pages/alerts/containers/alerts_page/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { WrappedAlertsPage as AlertsPage } from './alerts_page';
+export type { TopAlert } from './alerts_page';
