+++ x-pack/plugins/observability/public/pages/alerts/containers/alerts_page/styles.scss
